mark the virtual shelly switches as optimistic
esphome >= 2021.9.0 requires switches to either be optimistic, or have
turn_on_action and turn_off_action defined.

the turn_*_actions are only triggered when the switch is hit via the
API [1]:

> The action that should be performed when the remote (like Home Assistant’s
> frontend) requests the switch to be turned on.

However, we set these as "internal: true" as we don't need them to be
exposed in HA at all, so we can also set "optimistic: true" and move on
with our lives.

Fixes: #6

[1] https://esphome.io/components/switch/template.html